+++ src/erlkaf_nif.erl
@@ -12,7 +12,7 @@
     producer_cleanup/1,
     producer_set_owner/2,
     producer_topic_new/3,
-    produce/6,
+    produce/7,
     get_metadata/1,
 
     consumer_new/4,
@@ -48,7 +48,7 @@ producer_set_owner(_ClientRef, _Pid) ->
 producer_topic_new(_ClientRef, _TopicName, _TopicConfig) ->
     ?NOT_LOADED.
 
-produce(_ClientRef, _TopicRef, _Partition, _Key, _Value, _Headers) ->
+produce(_ClientRef, _TopicRef, _Partition, _Key, _Value, _Headers, _Timestamp) ->
     ?NOT_LOADED.
 
 get_metadata(_ClientRef) ->