+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 #### v2.1.0
 
 - Fix for kafka tombstones messages (https://github.com/silviucpp/erlkaf/issues/51)
-
+- Add support for message timestamp (https://github.com/silviucpp/erlkaf/issues/41)
 
 #### v2.0.9